pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask==3.1.0",
    "flask-bcrypt==1.0.1",
    "gunicorn==23.0.0",
    "mysql-connector==2.2.9",
    "mysql-connector-python==9.1.0",
]
